.ipynb_checkpoints/requirements-checkpoint.txt
opencv-python
Pillow
 
